% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_function.R
\name{estimate_price}
\alias{estimate_price}
\title{Estimate the price through a multiple linear regression}
\usage{
estimate_price(rooms, meter_square, location, data)
}
\arguments{
\item{rooms}{The number of rooms.}

\item{meter_square}{The number of square meters.}

\item{location}{The location (postal code).}

\item{data}{The dataset containing information on rooms, meter_square, price, and location}
}
\value{
The estimated prices of apartments with confidence intervals.
}
\description{
Estimate the price through a multiple linear regression
}
\examples{
estimate_price(2, 100, 1000, lausanne)
}
