% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors_function.R
\name{find_nearest_neighbors}
\alias{find_nearest_neighbors}
\title{Find Nearest Neighbors based on rooms and square meters}
\usage{
find_nearest_neighbors(rooms, meter_square, location, data, k = 5)
}
\arguments{
\item{rooms}{Number of rooms specified by the user}

\item{meter_square}{Number of square meters specified by the user}

\item{location}{Location (postal code) specified by the user}

\item{data}{The dataset containing information on rooms, meter_square, price, and location}

\item{k}{Number of nearest neighbors to retrieve (default is 5)}
}
\value{
A data frame with the k nearest neighbors matching the query
}
\description{
Find Nearest Neighbors based on rooms and square meters
}
\examples{
find_nearest_neighbors(5, 210, 1004, lausanne, k = 10)
}
