% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lausanne.R
\docType{data}
\name{lausanne}
\alias{lausanne}
\title{Lausanne Dataset}
\format{
## lausanne
A data frame with columns:
\describe{
  \item{rooms}{Number of rooms.}
  \item{meter_square}{Number of square meters.}
  \item{location}{Location (postal code).}
  \item{price}{Apartment price.}
}
}
\source{
This dataset is based on real estate data from Lausanne scraped on the immoscout website.
}
\usage{
lausanne
}
\description{
This dataset contains information about apartments in Lausanne.
}
\keyword{datasets}
